+++ fpsgame/src/main.rs
@@ -1,27 +1,12 @@
-use bevy::{
-    app::AppExit, 
-    input::keyboard::{
-        KeyCode,
-    }, input::mouse::{
-        MouseMotion,
-        MouseButton,
-    }, prelude::*, render::camera::Camera, window::{
-        WindowMode,
-    }};
-use bevy_rapier3d::{
-    physics::RapierPhysicsPlugin, 
-    render::RapierRenderPlugin, 
-    rapier::{
-        geometry::ColliderBuilder, 
-        dynamics::RigidBodyBuilder
-    }
-};
+use bevy::{input::keyboard::KeyCode, input::mouse::MouseMotion, prelude::*, render::camera::Camera, window::WindowMode};
+use bevy_rapier3d::{physics::RapierPhysicsPlugin, render::RapierRenderPlugin, rapier::{geometry::ColliderBuilder, dynamics::RigidBodyBuilder}};
 
 use player::Player;
 
 mod player;
 mod physics;
 mod math;
+mod game_state;
 
 fn main() {
     App::build()
@@ -43,7 +28,7 @@ fn main() {
         .add_system(update_look_direction.system())
         .add_system(move_player.system())
         .add_system(update_camera.system())
-        .add_system(toggle_cursor_and_exit.system())
+        .add_system(game_state::toggle_cursor_and_exit.system())
         .run();
 }
 
@@ -153,37 +138,6 @@ fn update_camera(
         for (_camera, mut transform) in camera_query.iter_mut() {
             *transform = Transform::from_translation(player_transform.translation)
             .looking_at(player_transform.translation + direction, Vec3::unit_y());
-        }
-    }
-}
-
-pub fn toggle_cursor_and_exit(
-    keyboard_input: Res<Input<KeyCode>>,
-    mouse_button_input: Res<Input<MouseButton>>,
-    mut windows: ResMut<Windows>,
-    mut app_exit_events: ResMut<Events<AppExit>>,
-){
-    if let Some(window) = windows.get_primary_mut() {
-        if keyboard_input.just_pressed(KeyCode::Escape)
-        {
-            if window.cursor_locked()
-            {
-                //unlock the cursor if it's locked
-                window.set_cursor_lock_mode(false);
-                window.set_cursor_visibility(true);
-            }
-            else
-            {
-                //exit the app if the cursor is unlocked
-                app_exit_events.send(AppExit);
-            }
-        }
-
-        if mouse_button_input.just_pressed(MouseButton::Left) && 
-           !window.cursor_locked()
-        {
-            window.set_cursor_lock_mode(true);
-            window.set_cursor_visibility(false);
         }
     }
 }
